requirements.txt
numpy
scikit-learn
umap-learn
matplotlib
networkx
ete3
jupyter
ruamel.yaml
pandas